Rename to entity-page-utils since sorting is utilized on both read and edit
